Update views.md
Good day Mr. Karam,

This is my first time contributing to the efforts of translating the Laravel framework into Arabic. I would appreciate any input you have about my translation so far. I am relying on حسوب to get all the proper word translation for the documentation as per your suggestion. 

I have a few questions. How 'formal' should the Arabic used be? For example, there is a big difference between saying مصفوفة, which is in English 'Matrix' and صفيفة, which is the programming entity called 'Array'. Do we care for such nuance? What about avoiding common mistakes in Arabic? Such as avoiding the use of 'تم' whenever possible. As long as the translation is readable, anything is ok, correct?

Thanks again for your efforts.